BOT-119
Make menuHelpers external function
